Virtual painter
The Virtual Painter project is an interactive computer vision-based application that allows users to paint and draw virtually using color detection and gesture tracking techniques. The application utilizes OpenCV and NumPy libraries in Python to access the webcam feed, track hand movements, and detect the colors specified by the user.